+++ R/Renderer.R
@@ -25,6 +25,9 @@ Renderer <- R6::R6Class( # nolint: object_name_linter.
     #'
     #' @param blocks `list` of `c("TextBlock", "PictureBlock", "NewpageBlock")` objects.
     #' @param yaml_header `character` a `rmarkdown` `yaml` header.
+    #' @param global_knitr `list` a global knitr parameters, like echo.
+    #' But if local parameter is set it will have priority.
+    #' Defaults to empty `list()`.
     #' @return `character` a `Rmd` text (`yaml` header + body), ready to be rendered.
     #' @examples
     #' card1 <- teal.reporter:::ReportCard$new()
@@ -57,16 +60,24 @@ Renderer <- R6::R6Class( # nolint: object_name_linter.
     #'
     #' yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
     #' result_path <- teal.reporter:::Renderer$new()$renderRmd(reporter$get_blocks(), yaml_header)
-    renderRmd = function(blocks, yaml_header) {
+    renderRmd = function(blocks, yaml_header, global_knitr = list()) {
       checkmate::assert_list(blocks, c("TextBlock", "PictureBlock", "NewpageBlock", "TableBlock", "RcodeBlock"))
       if (missing(yaml_header)) {
         yaml_header <- md_header(yaml::as.yaml(list(title = "Report")))
       }
-
       parsed_yaml <- yaml_header
-      parsed_blocks <- paste(unlist(lapply(blocks, function(b) private$block2md(b))), collapse = "\n\n")
+      parsed_global_knitr <- sprintf(
+        "\n```{r setup, include=FALSE}\nknitr::opts_chunk$set(%s)\n```\n",
+        capture.output(dput(global_knitr))
+      )
+      parsed_blocks <- paste(
+        unlist(
+          lapply(blocks, function(b) private$block2md(b))
+        ),
+        collapse = "\n\n"
+      )
 
-      rmd_text <- paste0(parsed_yaml, "\n", parsed_blocks, "\n")
+      rmd_text <- paste0(parsed_yaml, "\n", parsed_global_knitr, "\n", parsed_blocks, "\n")
       tmp <- tempfile(fileext = ".Rmd")
       input_path <- file.path(
         private$output_dir,
@@ -79,6 +90,9 @@ Renderer <- R6::R6Class( # nolint: object_name_linter.
     #'
     #' @param blocks `list` of `c("TextBlock", "PictureBlock", "NewpageBlock")` objects.
     #' @param yaml_header `character` an `rmarkdown` `yaml` header.
+    #' @param global_knitr `list` a global knitr parameters, like echo.
+    #' But if local parameter is set it will have priority.
+    #' Defaults to empty `list()`.
     #' @param ... `rmarkdown::render` arguments, `input` and `output_dir` should not be updated.
     #' @return `character` path to the output
     #' @examples
@@ -110,9 +124,9 @@ Renderer <- R6::R6Class( # nolint: object_name_linter.
     #'
     #' yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
     #' result_path <- teal.reporter:::Renderer$new()$render(reporter$get_blocks(), yaml_header)
-    render = function(blocks, yaml_header, ...) {
+    render = function(blocks, yaml_header, global_knitr = list(), ...) {
       args <- list(...)
-      input_path <- self$renderRmd(blocks, yaml_header)
+      input_path <- self$renderRmd(blocks, yaml_header, global_knitr)
       args <- append(args, list(
         input = input_path,
         output_dir = private$output_dir,
@@ -149,6 +163,7 @@ Renderer <- R6::R6Class( # nolint: object_name_linter.
     },
     rcodeBlock2md = function(block) {
       params <- block$get_params()
+      params <- lapply(params, function(l) if (is.character(l)) shQuote(l) else l)
       block_content <- block$get_content()
       sprintf(
         "\n```{r, %s}\n%s\n```\n",